fix: remove CRC, fix Viterbi crash, update interleaver

+++ include/ip/fec_codec.hpp
@@ -11,8 +11,8 @@ struct TransitionTarget {
 
 std::vector<uint32_t> hamming_encoder(const std::vector<uint8_t> &bytes);
 std::vector<uint8_t> hamming_decoder(const std::vector<uint32_t> &encoded_bytes);
-std::vector<uint32_t> interleaving(const std::vector<uint32_t> &input);
-std::vector<uint32_t> deinterleaving(const std::vector<uint32_t> &input);
+std::vector<uint8_t> interleaving(const std::vector<uint8_t> &input);
+std::vector<uint8_t> deinterleaving(const std::vector<uint8_t> &input);
 
 std::vector<uint8_t> conv_encoder(const std::vector<uint8_t> &bytes);
 std::vector<uint8_t> viterbi_decoder(const std::vector<uint8_t> &bytes);
